+++ 0.56.b/graysvr/CClientDialog.cpp
@@ -549,7 +549,7 @@ bool CDialogDef::GumpSetup( int iPage, CClient * pClient, CObjBase * pObjSrc, LP
 		return( false );
 
 	// starting x,y location.
-	int iSizes[2];
+	INT64 iSizes[2];
 	TCHAR * pszBuf = s.GetKeyBuffer();
 	m_pObj->ParseText( pszBuf, pClient->GetChar() );
 
@@ -706,7 +706,7 @@ TRIGRET_TYPE CClient::Dialog_OnButton( RESOURCE_ID_BASE rid, DWORD dwButtonID, C
 	// m_tmInpVal.m_PrvGumpID	= rid;
 	// m_tmInpVal.m_UID		= pObj ? pObj->GetUID() : (CGrayUID) 0;
 
-	int piCmd[3];
+	INT64 piCmd[3];
 	while ( s.ReadKeyParse())
 	{
 		if ( ! s.IsKeyHead( "ON", 2 ))
