+++ 0.56.b/graysvr/CItemStone.cpp
@@ -1438,7 +1438,7 @@ bool CItemStone::r_Verb( CScript & s, CTextConsole * pSrc ) // Execute command f
 			{
 				if ( s.HasArgs() )
 				{
-					int piCmd[2];
+					INT64 piCmd[2];
 					size_t iArgQty = Str_ParseCmds( s.GetArgStr(), piCmd, COUNTOF(piCmd));
 					if ( iArgQty == 2 )
 					{
